+++ spec/dummy_app/app/models/fruit.rb
@@ -5,4 +5,8 @@ class Fruit < ApplicationRecord
   if ENV["DB"] == "postgres"
     has_paper_trail versions: { class_name: "JsonVersion" }
   end
+
+  if PaperTrail.active_record_gte_7_0?
+    encrypts :supplier
+  end
 end