+++ CHANGELOG.md
@@ -7,7 +7,9 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Breaking Changes
 
-- None
+- [#1411](https://github.com/paper-trail-gem/paper_trail/pull/1411) - Change the return
+  value of #save_with_version when PaperTrail is disabled. It now returns the result
+  of calling #save on the model.
 
 ### Added